requirements.txt
django
django_rest_framework
django_allauth
django_rest_framework
django-rest-auth[with_social]
social-auth-app-django
google-cloud-translate
trimesh
pyglet
uwsgi
https://github.com/mercadopago/sdk-python.git#egg=mercadopago
django-mercadopago
psycopg2
psycopg2-binary
matplotlib
Pillow
django-imagekit
google-cloud-vision
